+++ src/accessories/simplicam.js
@@ -95,12 +95,16 @@ class SS3SimpliCam {
     startListening() {
         this.log(this.name + ' camera listening to alarm events...');
         this.simplisafe.subscribeToEvents((event, data) => {
+            if (!this.accessory) {
+                // Camera is not yet initialized
+                return;
+            }
             this.log(this.name + ` camera received new event from alarm: ${event}`);
             let eventCameraId;
             if (data && (data.sensorSerial || data.internal)) {
                 eventCameraId = data.sensorSerial ? data.sensorSerial : data.internal.mainCamera;
             }
-            
+
             switch (event) {
                 case 'CAMERA_MOTION':
                     if (eventCameraId == this.id) {
@@ -160,7 +164,8 @@ class CameraSource {
                     [480, 270, fps],
                     [640, 480, fps],
                     [640, 360, fps],
-                    [1280, 720, fps]
+                    [1280, 720, fps],
+                    [1920, 1080, fps]
                 ],
                 codec: {
                     profiles: [0, 1, 2],
@@ -177,6 +182,10 @@ class CameraSource {
             }
         };
 
+        let resolution = cameraConfig.cameraSettings.pictureQuality;
+        let maxSupportedHeight = +(resolution.split('p')[0]);
+        this.options.video.resolutions = this.options.video.resolutions.filter(r => r[1] <= maxSupportedHeight);
+
         this.createStreamControllers(2, this.options);
     }
 
@@ -186,9 +195,87 @@ class CameraSource {
         });
     }
 
-    handleSnapshotRequest(request, callback) {
-        this.log('Snapshot request. Not yet supported');
-        callback(new Error('Snapshots not yet supported'));
+    async handleSnapshotRequest(request, callback) {
+        let ffmpegPath = ffmpeg.path;
+        if (this.cameraOptions && this.cameraOptions.ffmpegPath) {
+            ffmpegPath = this.cameraOptions.ffmpegPath;
+        }
+        let resolution = `${request.width}x${request.height}`;
+        this.log(`Handling snapshot for ${this.cameraConfig.cameraSettings.cameraName} at ${resolution}`);
+
+        if (this.cameraConfig.model == 'SS001') { // Model(s) with privacy shutter
+            // Because if privacy shutter is closed we dont want snapshots triggering it to open
+            let alarmState = await this.simplisafe.getAlarmState();
+            switch (alarmState) {
+                case 'OFF':
+                    if (this.cameraConfig.cameraSettings.shutterOff !== 'open') {
+                        this.log(`SnapshotRequest ignored, ${this.cameraConfig.cameraSettings.cameraName} privacy shutter closed`);
+                        callback(new Error('Privacy shutter closed'));
+                        return;
+                    }
+                    break;
+
+                case 'HOME':
+                    if (this.cameraConfig.cameraSettings.shutterHome !== 'open') {
+                        this.log(`SnapshotRequest ignored, ${this.cameraConfig.cameraSettings.cameraName} privacy shutter closed`);
+                        callback(new Error('Privacy shutter closed'));
+                        return;
+                    }
+                    break;
+
+                case 'AWAY':
+                    if (this.cameraConfig.cameraSettings.shutterAway !== 'open') {
+                        this.log(`SnapshotRequest ignored, ${this.cameraConfig.cameraSettings.cameraName} privacy shutter closed`);
+                        callback(new Error('Privacy shutter closed'));
+                        return;
+                    }
+                    break;
+            }
+        }
+
+        try {
+            let newIpAddress = await dnsLookup('media.simplisafe.com');
+            this.serverIpAddress = newIpAddress.address;
+        } catch (err) {
+            if (!this.serverIpAddress) {
+                callback(new Error('Could not resolve hostname for media.simplisafe.com'));
+                return;
+            }
+        }
+
+        let sourceArgs = [
+            ['-re'],
+            ['-headers', `Authorization: Bearer ${this.simplisafe.token}`],
+            ['-i', `https://${this.serverIpAddress}/v1/${this.cameraConfig.uuid}/flv?x=${request.width}`],
+            ['-t', 1],
+            ['-s', resolution],
+            ['-f', 'image2'],
+            ['-vframes', 1],
+            ['-']
+        ];
+
+        let source = [].concat(...sourceArgs.map(arg => arg.map(a => typeof a == 'string' ? a.trim() : a)));
+
+        let ffmpegCmd = spawn(ffmpegPath, [
+            ...source,
+        ], {
+            env: process.env
+        });
+        this.log(ffmpegPath + source);
+
+        let imageBuffer = Buffer.alloc(0);
+
+        ffmpegCmd.stdout.on('data', data => {
+            imageBuffer = Buffer.concat([imageBuffer, data]);
+        });
+        ffmpegCmd.on('error', error => {
+            this.log('An error occurred while making snapshot request:', error);
+            callback(error);
+        });
+        ffmpegCmd.on('close', () => {
+            this.log(`Close ${this.cameraConfig.cameraSettings.cameraName} stream with image of length: ${imageBuffer.length}`);
+            callback(null, imageBuffer);
+        });
     }
 
     prepareStream(request, callback) {
@@ -259,8 +346,8 @@ class CameraSource {
             if (request.type == 'start') {
                 let sessionInfo = this.pendingSessions[sessionIdentifier];
                 if (sessionInfo) {
-                    let width = 1280;
-                    let height = 720;
+                    let width = 1920;
+                    let height = 1080;
                     let fps = this.cameraConfig.cameraSettings.admin.fps;
                     let videoBitrate = this.cameraConfig.cameraSettings.admin.bitRate;
                     let audioBitrate = 32;
@@ -393,29 +480,9 @@ class CameraSource {
                         }
                     }
 
-                    let source = [].concat(...sourceArgs.map(arg => arg.map(a => {
-                        if (typeof a == 'string') {
-                            return a.trim();
-                        } else {
-                            return a;
-                        }
-                    })));
-
-                    let video = [].concat(...videoArgs.map(arg => arg.map(a => {
-                        if (typeof a == 'string') {
-                            return a.trim();
-                        } else {
-                            return a;
-                        }
-                    })));
-
-                    let audio = [].concat(...audioArgs.map(arg => arg.map(a => {
-                        if (typeof a == 'string') {
-                            return a.trim();
-                        } else {
-                            return a;
-                        }
-                    })));
+                    let source = [].concat(...sourceArgs.map(arg => arg.map(a => typeof a == 'string' ? a.trim() : a)));
+                    let video = [].concat(...videoArgs.map(arg => arg.map(a => typeof a == 'string' ? a.trim() : a)));
+                    let audio = [].concat(...audioArgs.map(arg => arg.map(a => typeof a == 'string' ? a.trim() : a)));
 
                     let cmd = spawn(ffmpegPath, [
                         ...source,